requirements.txt
django<3
gunicorn
mysqlclient
PyGithub
requests
